requirements.txt
# Core Dependencies
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13
langchain-groq==0.2.2
langgraph==0.2.60

# API and Web Framework
fastapi==0.115.6
uvicorn==0.34.0
gunicorn==23.0.0
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
redis==5.2.0

# Vector Database & Embeddings (Phase 2)
chromadb==0.5.23
sentence-transformers==3.3.1
faiss-cpu==1.9.0

# UI
gradio==5.9.1

# Analytics and Visualization
plotly==5.24.1
pandas==2.2.2

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
aiofiles==23.2.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
flake8==7.1.1
black==24.10.0

# Logging and Monitoring
loguru==0.7.3

# Data Processing
numpy==1.26.4
